+++ src/views/Scripts.js
@@ -12,7 +12,7 @@ import {
   Col,
 } from "react-bootstrap";
 
-function Typography() {
+function Scripts() {
   return (
     <>
       <Container fluid>
@@ -144,4 +144,4 @@ function Typography() {
   );
 }
 
-export default Typography;
+export default Scripts;